+++ tests/lint-project/layers/App.Data.DAO.pas
@@ -1,0 +1,5 @@
+unit App.Data.DAO;
+interface
+uses System.SysUtils;
+implementation
+end.